missing repo and service test added

+++ api/src/pkg/services/responseDTO.go
@@ -1,9 +1,9 @@
 package services
 
 type NewCasesResponse struct {
-	Year       int     `json:"year"`
-	Month      int     `json:"month"`
-	Location   string  `json:"location"`
-	TotalCases float32 `json:"total_cases"`
-	NewCases   float32 `json:"new_cases"`
+	Year       int    `json:"year"`
+	Month      int    `json:"month"`
+	Location   string `json:"location"`
+	TotalCases int    `json:"total_cases"`
+	NewCases   int    `json:"new_cases"`
 }
